Changed task2 task4

+++ work/task2.h
@@ -1,4 +1,5 @@
 #define MAX 1000000
 
-unsigned int seqCollatz(unsigned int *maxlen);
-unsigned int collatz(unsigned  long num);
+
+unsigned long long seqCollatz(unsigned int *maxlen);
+unsigned int collatz(unsigned long long num);
